+++ src/ui/vanilla/entry-point/index.ts
@@ -3,7 +3,7 @@ import "../../../../libs/core/styles/spacings.css";
 import "../../../../libs/core/styles/shadows.css";
 
 import "./ConsoleWrapper.ts";
+import "./LivenessProve.ts";
 import "./MainSetup.ts";
-import "./MemoryLeakWorkaround.ts";
 import "./PreventDefaults.ts";
 import "./UxImprovements.ts";